[Q20-3] moved getToken methods to citeshere service
