+++ quadriga/src/main/java/edu/asu/diging/quadriga/api/v1/AddNetworkApiController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import edu.asu.diging.quadriga.api.v1.model.Quadruple;
 import edu.asu.diging.quadriga.api.v1.model.TokenInfo;
+import edu.asu.diging.quadriga.core.citesphere.impl.CitesphereConnectorImpl;
 import edu.asu.diging.quadriga.core.exception.NodeNotFoundException;
 import edu.asu.diging.quadriga.core.exceptions.OAuthException;
 import edu.asu.diging.quadriga.core.model.EventGraph;
@@ -23,7 +24,6 @@ import edu.asu.diging.quadriga.core.model.events.CreationEvent;
 import edu.asu.diging.quadriga.core.service.EventGraphService;
 import edu.asu.diging.quadriga.core.service.MappedTripleService;
 import edu.asu.diging.quadriga.core.service.NetworkMapper;
-import edu.asu.diging.quadriga.core.service.TokenValidator;
 
 @Controller
 public class AddNetworkApiController {
@@ -38,7 +38,7 @@ public class AddNetworkApiController {
     private MappedTripleService mappedTripleService;
     
     @Autowired
-    private TokenValidator tokenValidator;
+    private CitesphereConnectorImpl citesphereConnectorImpl;
 
     /**
      * The method parse given Json from the post request body and add Network
@@ -65,7 +65,7 @@ public class AddNetworkApiController {
         
         TokenInfo tokenInfo;
         try {
-            tokenInfo = tokenValidator.getTokenInfo(token);
+            tokenInfo = citesphereConnectorImpl.getTokenInfo(token);
             
             // either token info wasn't returned by citesphere or the token has expired
             if(tokenInfo == null || !tokenInfo.isActive()) {
