+++ quadriga/src/main/java/edu/asu/diging/quadriga/core/citesphere/CitesphereConnector.java
@@ -2,8 +2,11 @@ package edu.asu.diging.quadriga.core.citesphere;
 
 import java.util.List;
 
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.client.HttpClientErrorException;
 
+import edu.asu.diging.quadriga.api.v1.model.TokenInfo;
+import edu.asu.diging.quadriga.core.exceptions.OAuthException;
 import edu.asu.diging.quadriga.core.model.citesphere.CitesphereAppInfo;
 
 public interface CitesphereConnector {
@@ -13,5 +16,7 @@ public interface CitesphereConnector {
      * @return the list of apps
      */
     List<CitesphereAppInfo> getCitesphereApps() throws HttpClientErrorException;
+    
+    TokenInfo getTokenInfo(String token) throws BadCredentialsException, OAuthException;
 
 }
