--- quadriga/src/main/java/edu/asu/diging/quadriga/core/service/TokenValidator.java
@@ -1,12 +1,0 @@
-package edu.asu.diging.quadriga.core.service;
-
-import org.springframework.security.authentication.BadCredentialsException;
-
-import edu.asu.diging.quadriga.api.v1.model.TokenInfo;
-import edu.asu.diging.quadriga.core.exceptions.OAuthException;
-
-public interface TokenValidator {
-       
-    public TokenInfo getTokenInfo(String token) throws BadCredentialsException, OAuthException;
-    
-}
